Create artifact registry for docker from GitHub
Create a Docker repository in Artifact Registry, for GitHub Actions to
push built images into.

This includes a service account for GitHub Actions to impersonate, but
GitHub can't be given permission to imporsonate that account yet,
because its name isn't known.  A subsequent pull request will grant that
permission, along with a workflow and dockerfile.