Remove the Terraform GitHub Action
It's insecure, see https://stackoverflow.com/questions/64553739.

The way it affects us is that, for it to work at all, the GCP project
has to allow GitHub to use a service account that has "owner"
permissions at the project level (otherwise terraform wouldn't be able
to do very much).  The repo is private, so only GDSers can make pull
requests, but any GDSer can make a pull request, with a workflow in it
that uses terraform to make arbitrary changes.

We could restrict the repo to certain people or groups, but whoever we
allow write access to the repo would effectively also get owner
privileges on the project, via the Terraform GitHub Action.