Use authoritative IAM for the project
This makes explicit the IAM roles that are granted to principles at the
project level.

Most of them are associated with default service accounts that are
managed by Google.  Those shouldn't be interfered with, and when they
change, the terraform configuration and state should be synchronised
with GCP so that `terraform apply` does not make any changes.

That can be done by using `terraform import` and `terraform show` to
obtain and view the true state of the project, and then manually editing
the file `terraform/main-gcp.tf` to describe that state.  Once
`terraform plan` no longer proposes changes, it is safe to continue.