Add bank holidays
These are maintained in https://github.com/alphagov/frontend, but we
fetch them from the API, and parse the JSON in Neo4j cypher.

The query requires the APOC Core plugin for Neo4j, which is available in
the docker container, but isn't configured by default.  Now it is
configured, and a couple of other things are tidied up.