Configure postgres instance with local SSD
Another attempt to speed up restoring the Publishing API database, that
brought it down to 58 minutes.

On my laptop (16GB memory, 2 CPUs, local SSD) it took 54 minutes to
restore the whole database, indexes and all.  While one CPU restored the
`editions` table, the other CPU did all the other tasks, and finished
them ahead of the `editions` table, so 2 CPUs is enough.

What slows down the instance seems to be the network latency between it
and its disk.

Running the whole docker instance in memory nearly worked -- it ran for
over an hour, but I think had nearly finished, because it used 64GB of
disk, and it takes up 73GB on my laptop.  Unfortunately, only 64GB of
tmpfs was available, because tmpfs is half the amount of memory, and the
most that I asked for was 128GB.  Instances become quite expensive at
that point.

It's fiddly to mount a local SSD from inside a container, and the
documentation is poor.  The startup script has to be configured in a
different tag of the instance metadata, and mounting a directory is
undocumented.  In the configuration, I've linked to the clues that I
followed.

The startup script runs on the instance before the container is started.
It formats and mounts the local SSD, and creates a couple of folders
inside it.  The container mounts each of those folders, one for postgres
to use, and the other for the database backup file.  That leaves very
little for the instance to store itself, so its disk can be small.

The instance that is used is equivalent to my laptop configuration.  2
CPUs, 16GB memory, and a local SSD.