Extract roles from the Publishing API database
These are documents, the same as ones from the Content Store (mongodb),
but ones that don't have a base_path aren't in the Content Store as
documents in their own right.

As long as a person is in the Content Store who has an appointment to a
role, then the role is in the expanded_links, but it's best to go to the
Publishing API database to be certain.

Roles also have body text, but it is only available as govspeak
(GOV.UK's own flavour of markdown), so the govspeak ruby library is used
to render it to HTML, and then BeautifulSoup is used in Python scripts
to extract the plain text and any embedded hyperlinks (same as the
MongoDB data pipeline).

The data is uploaded into the govuk-knowledge-graph-data-processed
bucket, in a folder called publishing-api, so the filenames won't clash
with the ones from the Content Store (MongoDB) ... however they will
clash in the Neo4j import folder, so are prefixed with "role_".