Fix roles and link with persons
In the Publishing API, a `role_appointment` is a document that links to
documents of type `role` and `person`.  The role appointment has some
metadata (is it current? start date, optional end date).

In the AWS version of the knowledge graph, these relationships are
elided, so that a `Person` node has a direct `HAS_ROLE` relationship
with a `Role` node, and the start and end dates are properties of the
relationship.

This commit implements both versions.  Nodes are created for the
`role_appointment` document, with `LINKS_TO` relationships to a `Role`
and `Person` node.  Then a direct relationship `HAS_ROLE` is inferred,
with the start and end dates of the `role_appointment` document as its
properties. It omits the roleAppointmentContentId, which doesn't seem to
be required.

Reasons for implementing `role_appointment` and `LINKS_TO`:

- Consistency with the fact that other relationships are inferred from
  `LINKS_TO`.
- It's easier to infer the `HAS_ROLE` relationship in cypher than in
  SQL.
- RDF compatibility :)

This all takes quite a lot of code, in addition to all the
infrastructure for obtaining the Publishing API database in the first
place.  Perhaps we ought to use _only_ the Publishing API (postgres)
database, and not the Content Store (mongodb) database.