Create uniqueness constraints on all node types
This revealed a bug with `Organisation` and `Person` nodes, where
duplicates existed because all locales were used.  For now, only pages
in the `en` locale will be used to create `Organisation` and `Person`
nodes, but we are also considering changing the schema so that other
locales can also be used.

The queries that create relationships between Organisation nodes and
other nodes are now more explicit about node types, to ensure that
relationships aren't accidentally created between Page nodes instead.