Increase postgres shared memory
Increase the amount of shared memory available.
This requires the container to run in privileged mode.
It prevents a postgres error
"could not resize shared memory segment: No space left on device"

See https://stackoverflow.com/questions/56751565/pq-could-not-resize-shared-memory-segment-no-space-left-on-device

It can be done when the image is built, by using docker compose https://stackoverflow.com/questions/30210362/how-to-increase-the-size-of-the-dev-shm-in-docker-container

But then the image can't be tagged by a variable, so instead do it once
the image has booted (requires container to run in privileged mode)
https://stackoverflow.com/questions/66456581/gcc-vm-machine-allocate-more-dev-shm-memory

How much to configure?  More than `shared_buffers`, which is currently
4GB.
https://dba.stackexchange.com/questions/275378/dev-shm-size-recommendation-for-postgres-database-in-docker