Organise code by schema_name not document_type
Closes #196

The `schema_name` maps to how information in each document in the
content_store is laid out.  For example,
[`role`](https://github.com/alphagov/govuk-content-schemas/blob/main/dist/formats/role/frontend/schema.json)
is the schema of all role-related document types other than
[`role_appointment`](https://github.com/alphagov/govuk-content-schemas/blob/main/dist/formats/role_appointment/frontend/schema.json),
which is another schema.

This means not having to maintain the list of document types that is in
[`src/mongodb/js/body.js`](https://github.com/alphagov/govuk-knowledge-graph-gcp/blob/main/src/mongodb/js/body.js)
and other files.