Add Transaction as a new node type
Since transactions are marked as such in the content store
it's possible to make them first-class nodes in the Graph.
That would for instance make it possible to show an info-box
about the service when the user searched for it (e.g.
'universal credit' would show the description and direct link
in the info-box)