Load-test Neo4j for Gov Graph Search
This is a README and code to do some simple load-testing of Neo4j, to
check whether it can serve as many users of Gov Graph Search as we
expect.

- Define a complex example of a Gov Graph Search query
- Simulate 100 concurrent users running that query, on a running Neo4j
  instance

The result today was that a single user would be served in about 0.3
seconds, whereas a user among 100 concurrent users might have to wait
up to 3 seconds.  That is a tenfold increase in latency, but it isn't
important, and we don't expect so many concurrent users.