Use the 'english' analyzer for our indexes
In order to be as close as possible to the CONTAIN behaviour we
need our indexes not to do stemming. The 'cypher' analyser would
be ideal but doesn't seem to be working (reported at [1]). Same problem
with the keyword analyzer. 'english' seems to be the index that returns
the best results but doesn't work well with foreign-language pages. We'll
experiment with GGS once 'english' is used here.

[1] https://github.com/alphagov/govuk-knowledge-graph-gcp/issues/287