Inherit more properties from parent
Similarly to publishingApp, those properties don't exist on
parts. So we copy them from the parent Page. updatedAt won't
always be accurate though, since all parts will be marked as
updated even if only one has actually been changed. And we're
missing taxons and organisations, which are a bit more complex
to attach so we leave them to a future PR