Consolidate content of all document types
- Concatenate the various *_content tables into a single table.
- Derive a *_lines table from the content table.

This makes the content more accessible, because it's within a single
table, instead of being in separate table for each set of document
schemas.  It hides an implementation detail from the user -- the code
needs to know that content is in different places in each document
schema, but the user shouldn't have to know that.