Add more gcp services to terraform
These services were been enabled by work that was ephemeral or that
hasn't yet been terraformed.  When they were enabled, GCP automatically
created certain service accounts that are in the IAM configuration, so
the services must also be created in the 'dev' terraform config, so that
it can mirror the prod config.  Hence they ought to be enabled in the
prod config too.