Delete partial files of exported data (close #330)
When BigQuery exports a big table, it exports it into several small
files. There's a `gcloud storage compose` step that concatenates these
files into one big file. It ought to also remove the small files,
otherwise, if the source table were to shrink, some of the files
wouldn't be overwritten or removed, and their data would still be
concatenated into the big file.