Prevent accidentally destroying GCP project
It's too easy to accidentally destroy a GCP project with terraform when
switching between prod, dev and staging, because if you accidentally
change the project name in one of the terraform folders, it will try to
rename the project by deleting it first, and then recreating it.  The
lifecycle rule prevents it from doing that.