Include root taxons of all hierarchies
There is more than one hierarchy of taxons.  The main one is rooted at
the GOV.UK homepage, but another major one is rooted at /world/all, and
there are a couple of other small and strange ones. Although we were
pulling in all taxons in the mongodb pipeline, we weren't getting them
in `taxon_levels`, because we were only tracing the hierarchy down from
the root `/`.

Closes #373