Include ancestors of all tagged taxons in page
The page table should include the titles of the ancestors of all the
taxons that the page is tagged to, when a page is tagged to multiple
taxons.

A test case is
https://www.gov.uk/government/publications/air-passenger-duty-rates,
which is tagged to both "Air passenger duty" and "Air Passenger Duty"
(note capitalisation).

The set of ancestors of both taxons is:

- Air Passenger Duty
- Air passenger duty
- Aviation
- Business tax
- GOV.UK homepage
- Money
- Transport