In-progress: refactor tables for govsearch
Too much SQL logic was in the GovGraph search app:
https://github.com/alphagov/govuk-knowledge-graph-search

This refactor moves as much logic as possible into scheduled queries in
BigQuery.

The 'page' and 'part' tables in the 'graph' dataset will no longer have
array columns that contain information about taxons, organisations and
hyperlinks.

Tables in the new `search` dataset will contain array columns that
contain information about related things.  Their schemas will match as
closely as possible the data types defined in the GovSearch app:
https://github.com/alphagov/govuk-knowledge-graph-search/blob/541671c97f023e719d492b1eb4330df46ac6fae5/src/ts/search-api-types.ts

The new tables in the `search` dataset are updated by scheduled queries
that currently run at 06:00 daily.  Because they are scheduled, they
mustn't depend on each other, so they only depend on tables in the
`content` and `graph` datasets.  If this becomes unmanageable, we could
execute them from a virtual machine instead, and use a makefile or DBT
to manage the dependencies.