Handle travel advice summary parts in Neo4j
We now extract the 'summary' part of travel advice pages correctly.
However, because they have the same URL as the overall page, Neo4j
encounters the URL in both url.csv and parts.csv.  That means that we
have to use `MERGE`, not `CREATE` when creating nodes from parts.csv, so
that it reuses any node already created from url.csv.

To speed up the `MERGE`, it's worth creating the uniqueness constraint
before loading parts.csv.