Terraform the new gov-search.service.gov.uk domain
This required abandoning the convenient, official Google terraform
module for load balancers, and replacing it with several terraform
resources written by hand.  Doing so avoids down-time when adding a
domain to the certificate (which is what the module would do, replacing
the certificate altogether), by instead adding a certificate to the
https proxy.

https://cloud.google.com/blog/topics/developers-practitioners/new-terraform-module-serverless-load-balancing

https://github.com/terraform-google-modules/terraform-google-lb-http/issues/241