Fix govsearch domain terraform config (fix #418)
In #418 some resources in the dev and staging environments depended on a
variable that didn't exist.  The resources were a DNS record set and a
DNS managed zone.  This PR removes those resources from dev and staging.
In production, the resources are moved into environment.tf because they
are specific to that environment.

The variable definitions are reinstated in dev and staging, because they
have to be declared both in the .tfvars file and in main-gcp.tf.  We
could have removed them from main-gcp.tf, which would mean that
main-gcp.tf is yet another special file that varies between
environments.  Instead, we define them in the .tfvars file with a
default value of `"NOT_IN_USE"`.