Don't overwrite bank holiday raw table
A table that is configured in terraform, with metadata such as field
descriptions, is also being ovewritten by a daily workflow, which wipes
the field descriptions.  This creates noise in terraform plans, to
restore the field descriptions.  The fix is not to overwrite the table,
but instead to empty and then re-populate it.