Split bigquery terraform config into smaller files
The `bigquery.tf` file is unwieldy.  There is already a separate
`bigquery-search.tf` file to configure the `search` dataset.  This
commit extracts the configuration of the `content` and `graph` datasets
into separate files.  The `bigquery.tf` file remains to configure
miscellaneous things.

Nothing needs to be deployed after this commit, because nothing has
changed in the configuration itself.