Clean text on upload into Neo4j
GOV.UK content includes non-printable characters such as non-breaking
spaces.  In BigQuery, the `CONTAINS_SUBSTR()` function normalizes the
text before looking for matches, but Neo4j doesn't, so Neo4j searches
miss some matches that we believe users expect.  The `apoc.text.clean()`
function in Neo4j does the equivalent normalization to BigQuery, but it
takes a long time, so instead of asking users to use that function, this
commit uses the function when the content is imported into Neo4j.