Use page view statistics instead of pagerank
Users are confused about pagerank and assume that the statistic is
derived from the number of page views.  We think that the number of page
views will be just as useful for ranking search results.

This also removes a dependency on Neo4j to calculate the pagerank,
although #411 has some skeleton code to do calculate it in Python
instead.

The pagerank statistic remains, for now, because it does little harm.

The page view counts are loaded into Neo4j, because users still use
Neo4j.

The implementation is rather more complicated than for the similar task
of extracting page-to-page transition counts from GA4.  That task was
never tidy:

- BigQuery won't export a table to a single file
- `gcloud compose` can only compose a maximum of 32 files at once

We were relying on the export always fitting into exactly one file, by
chance.  Since then, @exfalsoquodlibet has developed a method with GCP
Workflows to recursively compose an arbitrary number of files.  It's a
faff, but it's robust, so we'll use it.