Abort postgres data pipeline if database too small
A problem with GOV.UK's backups means that the postgres database backup
file is corrupt, and too small.  The fact that the table exists triggers
the postgres pipeline, which wipes the BigQuery tables and is then
unable to repopulate them.  This commit checks the size of the database
file first, and aborts if it is too small.