Revert "Create table of links for the GovSearch app"
This reverts commit 2098a69a1e8920c4fe77b39ef8e54462457a1955 and pull
request #442.

When we came to implement the front-end search for links, we realised
that it would be necessary to filter links by the same filters that
apply to pages.  Including those filters would make the `search.link`
table almost the same as `search.page`, except that it would have one
row per page, per link, which would make the table much bigger than
`search.page` and therefore slower and more expensive to query.  The
only information that `search.link` adds to `search.page` is the page
title of destinations of links that are GOV.UK pages, and those titles
are only a nice-to-have, so we'll omit them for now and use the existing
`search.page` table.