Standardise IAP and Signon across environments
This is an alternative to
https://github.com/alphagov/govuk-knowledge-graph-gcp/pull/456.

Currently terraform/govgraphsearch.tf is different in the dev
environment from the other environments.  Ideally it would be the same,
so that any differences between environments are isolated in the
environment.auto.tfvars files, which is our convention.

Hypothesis: it's possible for both IAP and Signon to be configured at
the same time, without IAP interfering with Signon, as long as IAP
allows access to `allUsers`.  See
https://cloud.google.com/iap/docs/managing-access#public_access.  This
would mean that the difference between environments could be isolated to
environment.auto.tfvars, which lists the identities allowed by IAP.

Benefits:
- Migrate changes between environments more easily, by knowing that only
  the environment.auto.tfvars files vary.
- Debug access to the apps more easily, because every environment is
  similar, so tests can be done in non-production environments.
- Be ready to switch from controlling access via IAP to controlling it
  via Signon by changing only two variables in environment.auto.tfvars:
  `enable_auth` and `govgraphsearch_iap_members`.

Todo:
- Finish aligning the environments: only dev and prod have been aligned,
  staging has not.
- Check the hypothesis.