Remove the pagerank statistic
Updates of the pagerank table have been failing for a while.  Rather
than spend time fixing the updates of the table, we will stop
calculating the pagerank statistic.

- We no longer use the pagerank statistic in the GovSearch app.
- There are few users of Neo4j who would use it there.
- A reasonable alternative statistic to use is the number of page views.

This means we no longer need the Graph Data Science plugin in Neo4j.

There might be some firewall rules that are no longer required, which
allow the MongoDB instance to communicate with Neo4j, but there is a
separate issue to review all the firewall rules, so we won't edit them
here.