Adapt to changed schema of travel advice pages
The 'summary' section of travel advice pages was a fiddle, because it
wasn't among the other sections in the 'details.parts' field. That has
been corrected by alphagov/government-frontend#2747 so that travel
advice pages now have the same schema (in this regard) as guide pages.

This commit more or less reverts
2c51bb4aa2ce9098d9a8c57e8dc931d7472b4625, which was a bugfix to handle
the summary section in the first place.

Closes #457