Terraform IAM policies of Signon secrets
It turned out that these had been configured in the dev project, but not
terraformed, so that the Cloud Run service in the other projects
couldn't access the secrets.

This commit moves the definition of the service account to the top of
the file, because it is required by the new IAM configuration further
down, and because someone might need to set up this config step by step,
as explained in the existing comments.