Make terraform consistent in all environments
This adds one line to one file in one environment.  It enables an API
that is already enabled anyway, because terraform doesn't disable APIs
when they are removed from the configuration as this line was.  So there
will be nothing to deploy, and no actual real-world change except that
the `diff-terraform` GitHub Action should now pass.