Allow slug column to be null in graph.page
We mistakenly required the "slug" column to have a value in the
graph.page table, to be consistent with the content.parts table.  In
fact, the graph.page table includes rows of pages that aren't a "part"
of a guide document, and therefore don't have a slug.  Disallowing nulls
caused the query that populates the table to fail.