Consolidate some bigquery dataViewer roles
Currently we assign the bigquery.dataViewer to three different Google
Groups, each of which is managed independently of GovSearch.  It would
be better to create a single Google Group called govsearch-data-viewers,
and make those other groups members of govsearch-data-viewers.  Such a
group has been created, and this commit reconfigures terraform to use
it.