Parse content tables from json
This change implements a new structure to JSON schema files in the schema folder such that files now contain table details, including description, name, and friendly name.

This allows for a single terraform resource to loop through all of the schema files in the directory and create the required tables in a single step.

The possible advantages of this method are:

1. Terraform config becomes much simpler
2. Table definitions remain relatively clear in the separate .JSON files

The possible disadvantages of this method are:

1. JSON files are not in a prescribed/"official" format; the format would need to be documented
2. Each of the target tables would be a component of a single resource; this might make targeted changes to single tables more difficult down the line
3. The overall configuration has not been much simplified; the table configurations have simply been moved from .TF to .json files
4. It's not yet been tested/proved to work (pending!)