Configure plausible gov-search domain in tfvars
These domains don't exist (they would have to be delegated to us by
GOV.UK), but by being plausible they enable the creation of Cloud DNS
infrastructure in non-prod environments that match prod.  That, in turn,
makes terraform easier to manage, because more of it is the same in
every environment.