Add link text to the `hyperlinks` struct
It may be necessary to enable the population of link text instead of / along side link URL values for link searches.

The purpose of this change is to populate link text alongside link URLs (where link text is available).

Specifically, this change will:

1. Modify the schema of `search.page` table to include an additional field (`link_text`) in the `hyperlinks` struct;
2. Modify the `page.sql` SQL to populate the `link_text` field where possible (e.g. for embedded links).

It is not expected that this will necessitate front-end change in the meantime, where

1. The front-end already reads data from an array of records (the fundamental structure of the `search.page` table is not changing); and
2. The front-end has recently been made to read specific fields from the record values in the `hyperlinks` field.