Unable searching of text "snippets"
The purpose of this change is to enable searching through individual lines of content.

It is proposed that this be enabled by:

1. Updating the schema for `search.page` to replace the **string** "text" field with an array of **records**, containing the same content separated into one or more string "lines" (accompanied by the associated integer line number)
2. Updating the SQL populating the `search.page` table to populate the "text" field with content lines from the `content.lines` table.

The change assumes that it is required to populate the `line_number` alongside the `line` (content) on the grounds that:

 - It will only make a small difference in how the front-end will need to be modified to facilitate the change; and
 - It could be useful later on

Note that one side-effect of this change is to increase the volume of data processed in the query which populates `search.page` by somewhere nearing 1GB (was previously ~5.8GB, is now 6.8GB).

**Important**: implementing this change will necessitate a change in the [GovSearch (front-end) repository](https://github.com/alphagov/govuk-knowledge-graph-search).

It's likely that **at least** the [line 14](https://github.com/alphagov/govuk-knowledge-graph-search/blob/59560c096474843062e22e3637a087eca9fa9a86/src/backend/bigquery/buildSqlQuery.ts#L14)  in that repository will need to be amended to specify that the search should take place against field `text.line` (as opposed to just `text`).