Terraform the service account for CI/CD
This service account is already in use by the GitHub Actions of the
repository https://github.com/alphagov/govuk-knowledge-graph-search

It needs the roles/permissions that are documented here:
https://cloud.google.com/run/docs/reference/iam/roles#additional-configuration

This PR also addresses #405 by replacing the default bucket with one of
the same name, but in our preferred region.  To deploy this change,
manually delete the default bucket, and then apply the terraform to
recreate it.