Use service account blocks rather than strings
Assign members to roles by mentioning the terraform block rather than a
string.  This means that terraform can follow the dependencies between
IAM policies and service accounts, and avoid situations where terraform
deletes the entire policy just because one of the service accounts
doesn't exist yet.  We have locked ourselves out of the project a couple
of times now by falling into that trap.

Terraform should have an empty plan for this change.  Currently the plan
isn't empty, because there are discrepancies between the real
infrastructure and the configuration.