Include metadata in 'part' pages
GovSearch relies on "parts" of guide documents inheriting the same
metadata as the overall document, especially the publishing
organisations.  Otherwise, when a user filters by publishing
organisation, none of the parts appears in the results, because none of
them has a value for publishing organisation.

We can do this by linking via `content_id` (which is the same for each
part of a document) instead of `url` (which we construct to be different
for each part of a document).

We would be able to do this more simply if we hadn't unwisely chosen to
construct the tables `has_organisation` and
`has_primary_publishing_organisation` to use our own Frankenstein URL, a
concatenation of 'https://www.gov.uk/' with the content_id.  This PR
chooses not to use those tables.  Alternatively it could have used those
tables, and joined them by an expression that constructs a similar URL
for the page being joined to.