Time out govspeak rendering
Govspeak can freeze when rendering a string. In such a case, it is
better to return an error for that particular string after a short time
than to wait a long time for govspeak to fail the whole batch.

Example of freezing:

```ruby
govspeak = <<-END_GOVSPEAK
$LegislativeList
* Item
  $EndLegislativeList
END_GOVSPEAK
Govspeak::Document.new(govspeak).to_html
```