feat: BigQuery UDF to find phone numbers in text
This uses a Google library called libphonenumber to extract phone
numbers from plain text, and standardise them so that they are easier to
match.  For example "01234 567 890" might be standardised to
"+441234567890".  A user query in GovSearch could also be standardised,
so that comparisons are always between standard forms. Otherwise a user
might search for "01234 567890", which wouldn't match "01234 567 890".

Google's own javascript implementation is difficult to deploy, so we use
a different distribution that Google themselves suggest.  This is a
single javascript file that is stored in a bucket.  The "max" version is
used because GOV.UK has numbers in forms that aren't supported by the
other versions.

One UDF (user-defined-function) fetches the library from the bucket,
imports it into a short javascript to process some text, and returns the
results as JSON.

A second UDF wraps the first one, and parses the JSON into a STRUCT,
which is easier to work with.  This second function is necessary because
Javascript UDFs can't return structs themselves.