feat: incrementally derive current editions
The publishing database is huge, because it contains every version of
every document. We only need the current version of each document. We
could derive that straightforwardly (and even quite quickly) from
scratch every time, but it would cost hundreds of gigabytes each time,
so instead we do this incrementally by:

- Partitiioning tables so that BigQuery won't charge for the whole
  thing when only part of the table is relevant.
- Processing only the new records each time.