feat: remove unused tables in the 'search' dataset
Once the GovSearch infobox is removed by
https://github.com/alphagov/govuk-knowledge-graph-search/pull/230 (if
that happens), several tables in the `search` dataset can be removed.

Ones that must remain are `page` and `taxon`. The `taxon` table could
possibly be refactored.  New `organisation` and `locale` tables could be
created to populate drop-down menus, which are currently populated from
tables in other datasets.