feat: tables to populate govsearch dropdowns
Corresponds to https://github.com/alphagov/govuk-knowledge-graph-search/issues/63

It will save money by allowing GovSearch to query for exactly the data
that it needs, rather than filtering bigger datasets, which it otherwise
does for every new session.

The queries are also adapted to run from the `content` dataset (which is
raw data), rather than the `graph` dataset (which is derived data), so
that the `graph` dataset can eventually be removed.