feat: bigquery datasets - private and public
[Trello](https://trello.com/c/mlKACiMO/112-govgraph-make-some-data-open)

Create two new BigQuery datasets, called 'private' and 'public', for
data that doesn't belong elsewhere.

An example of data that belongs in the 'private' dataset is data that is
derived from the Publishing API data in the 'publishing-api' dataset,
and that must not be publicly available.

An example of data that belongs in the 'public' dataset is data that is
derived from the Content API data in the 'content-api' dataset,
and that may be publicly available.

An example of data that doesn't belong in either the 'public' or
'private' datasets is data that is used directly by the GovSearch app.
It is better for that data to be in its own dataset, so that the
GovSearch app can be granted access to only the data that it needs, and
no more.