refactor: add an intermediate step before public
New, current, online editions aren't just appended to a table. They are
also processed further, to extract the content. That means that it is
useful to keep an intermediate step between scrubbing them of non-public
data, and appending them to the main table.  That way, further
processing can use the small intermediate step, rather than
re-processing the whole main table.