refactor: use pandoc not selenium to extract text
Selenium + chromedriver is proving to be incredibly slow to extract
plain text from HTML. It is a pity, because it's better to extract text
as a widely-used browser renders it, than as an esoteric HTML parsing
library does. But for expediency, we'll try using pandoc, which is at
least more widely known than the w3m browser.