refactor: derive search tables from Publishing API
This is a step in deriving all the tables in the `search` dataset from
the Publishing API data.

A new batch query is defined and scheduled.

Routines are defined, to refresh each table, and they are called from
the batch query.

The taxon table schema is much reduced to the single column that the
GovSearch app uses.