fix: bugs in content extraction
It was assuming the wrong data types for some JSON elements.

It was also possible to simplify the handling of govspeak vs HTML.
Everything is extracted as though it is govspeak, even when it is HTML.
Then the 'govspeak' is rendered to HTML, which doesn't harm what was
already HTML.

More of the special cases could be supported than just guide,
travel_advice and step_by_step_nav. For example, hmrc_manual could be
supported better. But nobody has yet asked for it.