chore: purge the old Content API data pipeline
Everything is now derived from the Publishing API.

Remove the 'content_api' dataset. Leave intact the datasets 'content'
and 'graph', which are described as 'depecated', and their few remaining
tables are derived from the Publishing API data in the 'private' and
'public' datasets.

Remove the virtual machines that handled the PostgreSQL and MongoDB
versions of the Content API backup database.

Remove parts of the workflow that orchestrates the data pipeline.

Remove service accounts.

Remove GitHub Actions that built docker images.