feat: lookup table of real-world base paths
Create a table that maps base paths used in the real world to base paths
that belong to editions in the Publishing API database.  The difference
is due to documents of schema 'guide' and 'travel_advice' having a
real-world base_path for each part of the document.

This depends on the table public.content, which extracts the extra slugs
of 'guide' and 'travel_advice' documents.  The public.content table
can't itself be used as a lookup table, because it omits documents that
don't have content, or whose content we don't yet extract.