fix: Use main page of multiparts in GovSearch
Omit the first 'part' of multi-part documents, which duplicates (and is
currently redirected to) the main document that has no slug. For example,
include the following:

  /main-page
  /main-page/second-part
  /main-page/third-part

Omit the following:

  /main-page/first-part

The reason to include /main-page instead of /main-page/first-part is that
no page views are recorded for /main-page/first-part.

Currently, /main-page is excluded, and /main-page/first-part is
included. This commit chooses the alternative.