fix: support a new column in Publishing API table
The editions table in the upstream Publishing API database has a new
column, `last_edited_by_editor_id`, created in
https://github.com/alphagov/publishing-api/pull/2919. This broke the
data pipeline, because the BigQuery schema of that table rejected the
new column.

The new column contains sensitive data, so it is only included in tables
in the `publishing-api` and `private` datasets, which most users cannot
access. It is explicitly omitted from the query that puts the data into
more accessible datasets.