chore: upgrade MongoDB docker image for publisher
`apt-get update` now fails in the docker image of MongoDB version 4.4.16
due to an expired signing key. That version of MongoDB is no longer
supported, so the signing key is unlikely to be replaced. Upgrading to a
newer docker image solves this problem.

Version compatibility check:

* We only need the package `mongodb-database-tools`, which is versioned
  independently of MongoDB. The version of `mongodb-database-tools` that
  comes in the Docker image of the latest version of MongoDB is
  100.11.0.

* According to
  https://www.mongodb.com/docs/database-tools/mongorestore/mongorestore-compatibility-and-installation/#mongodb-server-compatibility
  `mongodb-database-tools` version 100.10.0 is compatible with MongoDB
  versions 5.0 and 4.0, so we presume that version 100.11.0 maintains
  that compatibility.

* The publisher database backup file is created by AWS DocumentDB, which
  according to
  https://docs.aws.amazon.com/documentdb/latest/developerguide/compatibility.html
  is compatible with MongoDB version 5.0 and 4.0.

Secondary problem: something to do with "Popular Links"
https://github.com/alphagov/publisher/pull/2235/files means that there
are four editions that have no slug, which caused a problem with
concatenation. There might still be a problem with assuming that they
are about the homepage, although they are.